Import issue solved, idle timeout -1, tutorial widget code clean up
